+++ scripts/05_depth_profiles/02_singlesub_depth_vs_T1.py
@@ -171,9 +171,9 @@ for s in SUBJ_ID:
         # Custom tick labels
         ax[i].set_xticks([0, 25, 49, 51, 100, 149, 151, 175, 200])
         ax[i].set_xticklabels([0.7, 0.35, None, 0, 0.5, 1, None, 0.35, 0.7])
-        ax[i].set_yticks(np.linspace(0, nr_bins-1, 6, dtype=np.int))
+        ax[i].set_yticks(np.linspace(0, nr_bins-1, 6, dtype=np.int32))
         ax[i].set_yticklabels(np.linspace(RANGE_Y[0], RANGE_Y[1], 6,
-                                          dtype=np.int))
+                                          dtype=np.int32))
 
         # Add text (positions are in data coordinates)
         ax[i].text(0 + 1, 80, 'Below\ngray matter\n(White matter)',
@@ -223,7 +223,7 @@ for s in SUBJ_ID:
 
         # # Set Y axis tick labels
         # ax[i].set_ylabel(YLABEL)
-        # ax[i].set_yticks(np.linspace(0, nr_bins-1, 6, dtype=np.int))
+        # ax[i].set_yticks(np.linspace(0, nr_bins-1, 6, dtype=np.int32))
         # yticks = np.linspace(RANGE_Y[0], RANGE_Y[1], 6) / 1000
         # yticks = ["{:.1f}".format(y) for y in yticks]
         # ax[i].set_yticklabels(yticks)
