+++ scripts/05_depth_profiles/04_group_depth_vs_T1.py
@@ -100,7 +100,7 @@ for i in range(4):
     # Custom tick labels
     ax[i].set_xticks([-0.7, -0.35, -0.01, 0.01, 0.5, 0.99, 1.01, 1.35, 1.7])
     ax[i].set_xticklabels([0.7, 0.35, None, 0, 0.5, 1, None, 0.35, 0.7])
-    yticks = np.linspace(RANGE_Y[0], RANGE_Y[1], 6, dtype=np.int)
+    yticks = np.linspace(RANGE_Y[0], RANGE_Y[1], 6, dtype=np.int32)
     ax[i].set_yticks(yticks)
     ax[i].set_yticklabels(yticks)
